cleaned up server
