+++ Planarian/Planarian/Modules/Account/Services/AccountService.cs
@@ -1,5 +1,4 @@
 using Planarian.Model.Shared;
-using Planarian.Modules.Account.Controller;
 using Planarian.Modules.Account.Repositories;
 using Planarian.Modules.Files.Repositories;
 using Planarian.Modules.Files.Services;
