+++ Planarian/Planarian/Modules/App/Repositories/AppRepository.cs
@@ -1,7 +1,5 @@
-using Microsoft.EntityFrameworkCore;
 using Planarian.Model.Database;
 using Planarian.Model.Shared;
-using Planarian.Modules.Authentication.Models;
 using Planarian.Shared.Base;
 
 namespace Planarian.Modules.App.Repositories;
