+++ Planarian/Planarian/Modules/Authentication/Services/AuthenticationService.cs
@@ -1,5 +1,4 @@
 using Planarian.Model.Shared;
-using Planarian.Model.Shared.Helpers;
 using Planarian.Modules.Authentication.Models;
 using Planarian.Modules.Authentication.Repositories;
 using Planarian.Modules.Users.Repositories;
