+++ Planarian/Planarian/Modules/Authentication/Services/TokenService.cs
@@ -1,7 +1,6 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
-using Microsoft.Extensions.Azure;
 using Microsoft.IdentityModel.Tokens;
 using Planarian.Library.Extensions.String;
 using Planarian.Library.Options;
