+++ Planarian/Planarian/Modules/Caves/Controllers/CaveController.cs
@@ -90,47 +90,4 @@ public class CaveController : PlanarianControllerBase<CaveService>
 
         return new OkResult();
     }
-
-    #region Import
-
-    [DisableRequestSizeLimit] //TODO
-    [HttpPost("import-caves/file")]
-    public async Task<IActionResult> ImportCavesFile(string? uuid, [FromForm] IFormFile file,
-        CancellationToken cancellationToken)
-    {
-        var result = await Service.AddTemporaryFileForImport(file.OpenReadStream(), file.FileName, uuid, cancellationToken);
-
-        return new JsonResult(result);
-    }
-    
-    [HttpPost("import-caves/process/{fileId:length(10)}")]
-    public async Task<IActionResult> ImportCavesFileProcess(string fileId,
-        CancellationToken cancellationToken)
-    {
-        var result = await Service.ImportCavesFileProcess(fileId, cancellationToken);
-
-        return new JsonResult(result);
-    }
-
-    [DisableRequestSizeLimit] //TODO
-    [HttpPost("import-entrances/file")]
-    public async Task<IActionResult> ImportEntrancesFile(string? uuid, [FromForm] IFormFile file,
-        CancellationToken cancellationToken)
-    {
-        var result =
-            await Service.AddTemporaryFileForImport(file.OpenReadStream(), file.FileName, uuid, cancellationToken);
-
-        return new JsonResult(result);
-    }
-
-    [HttpPost("import-entrances/process/{fileId:length(10)}")]
-    public async Task<IActionResult> ImportEntrancesFileProcess(string fileId,
-        CancellationToken cancellationToken)
-    {
-        var result = await Service.ImportEntrancesFileProcess(fileId, cancellationToken);
-
-        return new JsonResult(result);
-    }
-    
-    #endregion
 }
