+++ Planarian/Planarian/Modules/Files/Repositories/FileRepository.cs
@@ -1,7 +1,5 @@
-using System.Collections;
 using Microsoft.EntityFrameworkCore;
 using Planarian.Model.Database;
-using Planarian.Model.Database.Entities;
 using Planarian.Model.Shared;
 using Planarian.Modules.Files.Services;
 using Planarian.Shared.Base;
