+++ Planarian/Planarian/Modules/Files/Services/FileService.cs
@@ -1,10 +1,7 @@
 
 using System.ComponentModel.DataAnnotations;
-using Azure.Storage;
 using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
 using Azure.Storage.Sas;
-using Planarian.Model.Database.Entities.RidgeWalker;
 using Planarian.Model.Shared;
 using Planarian.Modules.Files.Controllers;
 using Planarian.Modules.Files.Repositories;
