+++ Planarian/Planarian/Modules/Import/Models/FailedCaveCsvRecord.cs
@@ -1,0 +1,15 @@
+namespace Planarian.Modules.Import.Models;
+
+public class FailedCaveCsvRecord<T>
+{
+    public FailedCaveCsvRecord(T rowData, int rowNumber, string reason)
+    {
+        CaveCsvModel = rowData;
+        RowNumber = rowNumber;
+        Reason = reason;
+    }
+
+    public T CaveCsvModel { get; set; }
+    public int RowNumber { get; set; }
+    public string Reason { get; set; }
+}
