+++ Planarian/Planarian/Modules/Trips/Repositories/TripRepository.cs
@@ -4,7 +4,6 @@ using Planarian.Model.Database.Entities;
 using Planarian.Model.Database.Entities.Leads;
 using Planarian.Model.Database.Entities.Trips;
 using Planarian.Model.Shared;
-using Planarian.Modules.Leads.Controllers;
 using Planarian.Modules.Photos.Models;
 using Planarian.Modules.Query.Extensions;
 using Planarian.Modules.Query.Models;
