+++ Planarian/Planarian/Program.cs
@@ -19,6 +19,7 @@ using Planarian.Modules.Caves.Repositories;
 using Planarian.Modules.Caves.Services;
 using Planarian.Modules.Files.Repositories;
 using Planarian.Modules.Files.Services;
+using Planarian.Modules.Import.Repositories;
 using Planarian.Modules.Leads.Repositories;
 using Planarian.Modules.Leads.Services;
 using Planarian.Modules.Notifications.Hubs;
