+++ Planarian/Planarian/Shared/Base/ServiceBase.cs
@@ -4,16 +4,20 @@ namespace Planarian.Shared.Base;
 
 public abstract class ServiceBase
 {
+    protected readonly RequestUser RequestUser;
+
+    protected ServiceBase(RequestUser requestUser)
+    {
+        RequestUser = requestUser;
+    }
 }
 
 public abstract class ServiceBase<TRepository> : ServiceBase where TRepository : RepositoryBase
 {
     protected readonly TRepository Repository;
-    protected readonly RequestUser RequestUser;
 
-    protected ServiceBase(TRepository repository, RequestUser requestUser)
+    protected ServiceBase(TRepository repository, RequestUser requestUser) : base(requestUser)
     {
         Repository = repository;
-        RequestUser = requestUser;
     }
 }
